+++ src/html5/input_stream.rs
@@ -1,4 +1,4 @@
-use crate::html5_parser::tokenizer::{CHAR_CR, CHAR_LF};
+use crate::html5::tokenizer::{CHAR_CR, CHAR_LF};
 use std::fs::File;
 use std::io::Read;
 use std::{fmt, io};
