+++ src/lib.rs
@@ -5,6 +5,6 @@ pub mod api;
 #[allow(dead_code)]
 pub mod css3;
 #[allow(dead_code)]
-pub mod html5_parser;
+pub mod html5;
 pub mod testing;
 pub mod types;
